# src/main/resources/static/documents/openapi.json
{"openapi":"3.0.1","info":{"title":"Security Example","description":"JWT Spring Security Implementation","contact":{"name":"rafinhaLQ","url":"https://github.com/rafinhaLQ","email":"rafinhalq@gmail.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080/api/let-go","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"Authentication API"},{"name":"User","description":"User API"}],"paths":{"/user/register":{"post":{"tags":["User"],"summary":"Register user","description":"Register user and return user id","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/authenticate":{"post":{"tags":["Authentication"],"summary":"Authenticate user","description":"Authenticate user and return JWT token","operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredencial"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtTokenData"}}}}}}},"/user/{user_id}":{"get":{"tags":["User"],"summary":"Get user info","description":"Get user info by user id","operationId":"getUserInfo","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"JWT":[]}]},"delete":{"tags":["User"],"summary":"Delete user","description":"Delete user by user id","operationId":"deleteUser","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"JWT":[]}]},"patch":{"tags":["User"],"summary":"Update user info","description":"Update user info by user id","operationId":"updateUser","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"JWT":[]}]}},"/user/{user_id}/roles":{"patch":{"tags":["User"],"summary":"Update user roles","description":"Update user roles by user id","operationId":"updateUserRoles","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"JWT":[]}]}},"/user/{user_id}/password":{"patch":{"tags":["User"],"summary":"Update user password","description":"Update user password by user id","operationId":"updateUserPassword","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserPasswordDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"JWT":[]}]}}},"components":{"schemas":{"CreateUserDto":{"required":["firstName","lastName","password"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"cep":{"pattern":"\\d{5}-?\\d{3}","type":"string"},"phoneNumber":{"pattern":"([(]?\\d{2}[)]?)\\d{5}-?\\d{4}","type":"string"}}},"UserCredencial":{"required":["password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"JwtTokenData":{"type":"object","properties":{"jwt_token":{"type":"string"}}},"CreateAddressDto":{"required":["cep"],"type":"object","properties":{"cep":{"pattern":"\\d{5}-?\\d{3}","type":"string"},"isMain":{"type":"boolean"}}},"UpdateUserDto":{"required":["addressDto","firstName","lastName"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"addressDto":{"$ref":"#/components/schemas/CreateAddressDto"},"phoneNumber":{"pattern":"([(]?\\d{2}[)]?)\\d{5}-?\\d{4}","type":"string"}}},"Address":{"type":"object","properties":{"cep":{"type":"string"},"addressLine1":{"type":"string"},"addressLine2":{"type":"string"},"district":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"main":{"type":"boolean"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"addresses":{"type":"array","items":{"$ref":"#/components/schemas/Address"}},"phoneNumber":{"type":"string"}}},"UpdateUserPasswordDto":{"required":["newPassword","oldPassword"],"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}}}},"securitySchemes":{"JWT":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}